AC-12 implement config builder
The config is now built from multiple files. Config files with apropriate
names are collected from the current directory up to the filesystem root.
All config files are then sorted by priority and finally merged.

Handling of an incomplete config was also improved: the user now gets feedback
which keys are missing along with the possiblity to fill them with default
values in the config file with the highest priority.